Revert "Merge branch 'master' into Drag-n-Drop"
This reverts commit 37f08b8f874edafcb94239066e79b641d5c8a796, reversing
changes made to 99a77192bb66566d459ef0cf1733a5e08f94d3d3.

+++ Assets/Scripts/UI/Pedido.cs
@@ -6,7 +6,7 @@ public class Pedido : MonoBehaviour
 {
     public float timer;
     public Pedido seg;
-    public int reward;
+    public float reward;
 
     public Pedido(Pedido seg = null)
     {
